+++ internal/transport/transport.go
@@ -160,7 +160,7 @@ type DefaultTransportFactory struct{}
 func (dtm *DefaultTransportFactory) Create(nhConfig config.NodeHostConfig,
 	handler raftio.MessageHandler,
 	chunkHandler raftio.ChunkHandler) raftio.ITransport {
-	return NewTCPTransport(nhConfig, handler, chunkHandler)
+	return NewUDPTransport(nhConfig, handler, chunkHandler)
 }
 
 // Validate returns a boolean value indicating whether the specified address is
